rename some files and folders
