+++ src/components/PaginationCommonPage/PaginationCommonPage.tsx
@@ -4,8 +4,11 @@ import {
   BlockButton,
   PaginationButtonStyle,
   PaginationButtonText,
-} from './PaginationControls.styles';
+} from './PaginationCommonPage.styles';
+import FetchStarShips from '../../utlis/FetchData/FetchStarShips';
+import FetchVehicles from '../../utlis/FetchData/FetchVehicles';
 import FetchCharacters from '../../utlis/FetchData/FetchCharacters';
+import {useNavigationState} from '@react-navigation/native';
 
 type paginationResponse = Pick<CommonTypes, 'next' | 'previous'>;
 
@@ -19,17 +22,32 @@ const initialState = {
   previous: '',
 };
 
-export default function PaginationControls({
+export default function PaginationCommonPage({
   currentPage,
   setCurrentPage,
 }: DirectionProps): JSX.Element {
   const [pagination, setPagination] =
     useState<paginationResponse>(initialState);
 
+  const title = useNavigationState(state => state.routes[state.index].name);
+
   useEffect(() => {
-    FetchCharacters(currentPage).then(({next, previous}: paginationResponse) =>
-      setPagination({next, previous}),
-    );
+    if (title === 'ScreenStarShips') {
+      FetchStarShips(currentPage).then(({next, previous}: paginationResponse) =>
+        setPagination({next, previous}),
+      );
+    }
+    if (title === 'ScreenVehicles') {
+      FetchVehicles(currentPage).then(({next, previous}: paginationResponse) =>
+        setPagination({next, previous}),
+      );
+    }
+    if (title === 'Home') {
+      FetchCharacters(currentPage).then(
+        ({next, previous}: paginationResponse) =>
+          setPagination({next, previous}),
+      );
+    }
   }, [currentPage]);
 
   const disabledButton = (
