+++ src/components/PaginationMainPage/PaginationMainPage.styles.tsx
