+++ src/components/PaginationMainPage/PaginationMainPage.tsx
@@ -4,10 +4,8 @@ import {
   BlockButton,
   PaginationButtonStyle,
   PaginationButtonText,
-} from './Pagination.styles';
-import FetchStarShips from '../../utlis/FetchData/FetchStarShips';
-import FetchVehicles from '../../utlis/FetchData/FetchVehicles';
-import {useNavigationState} from '@react-navigation/native';
+} from './PaginationMainPage.styles';
+import FetchCharacters from '../../utlis/FetchData/FetchCharacters';
 
 type paginationResponse = Pick<CommonTypes, 'next' | 'previous'>;
 
@@ -21,26 +19,17 @@ const initialState = {
   previous: '',
 };
 
-export default function Pagination({
+export default function PaginationMainPage({
   currentPage,
   setCurrentPage,
 }: DirectionProps): JSX.Element {
   const [pagination, setPagination] =
     useState<paginationResponse>(initialState);
 
-  const title = useNavigationState(state => state.routes[state.index].name);
-
   useEffect(() => {
-    if (title === 'ScreenStarShips') {
-      FetchStarShips(currentPage).then(({next, previous}: paginationResponse) =>
-        setPagination({next, previous}),
-      );
-    }
-    if (title === 'ScreenVehicles') {
-      FetchVehicles(currentPage).then(({next, previous}: paginationResponse) =>
-        setPagination({next, previous}),
-      );
-    }
+    FetchCharacters(currentPage).then(({next, previous}: paginationResponse) =>
+      setPagination({next, previous}),
+    );
   }, [currentPage]);
 
   const disabledButton = (
